Clean code & new features
